refactor: split non-SDK integration tests by behavior
