+++ tests/integration/src/rust_pipeline/functional/abi_transform/tx_kernel/mod.rs
@@ -1,0 +1,3 @@
+mod get_inputs;
+mod get_metadata;
+mod note_build_recipient;
