--- tests/integration/src/rust_pipeline/functional/regressions.rs
@@ -1,350 +1,0 @@
-use miden_core::Felt;
-use midenc_expect_test::expect_file;
-use midenc_frontend_wasm::WasmTranslationConfig;
-
-use crate::{
-    CompilerTest,
-    testing::{eval_package, setup},
-};
-
-#[test]
-fn test_func_arg_same() {
-    // This test reproduces the https://github.com/0xMiden/compiler/issues/606
-    let main_fn = r#"
-        (x: &mut Felt, y: &mut Felt) -> i32 {
-            intrinsic(x, y)
-        }
-
-        #[unsafe(no_mangle)]
-        #[inline(never)]
-        fn intrinsic(a: &mut Felt, b: &mut Felt) -> i32 {
-            unsafe { (a as *mut Felt) as i32 }
-        }
-
-    "#;
-
-    setup::enable_compiler_instrumentation();
-    let config = WasmTranslationConfig::default();
-    let mut test = CompilerTest::rust_fn_body_with_stdlib_sys("func_arg_same", main_fn, config, []);
-
-    let package = test.compile_package();
-
-    let addr1: u32 = 10 * 65536;
-    let addr2: u32 = 11 * 65536;
-
-    // Test 1: addr1 is passed as x and should be returned
-    let args1 = [Felt::from(addr1), Felt::from(addr2)];
-    eval_package::<i32, _, _>(&package, [], &args1, &test.session, |trace| {
-        let result: u32 = trace.parse_result().unwrap();
-        assert_eq!(result, addr1);
-        Ok(())
-    })
-    .unwrap();
-
-    // Test 1: addr2 is passed as x and should be returned
-    let args2 = [Felt::from(addr2), Felt::from(addr1)];
-    eval_package::<i32, _, _>(&package, [], &args2, &test.session, |trace| {
-        let result: u32 = trace.parse_result().unwrap();
-        assert_eq!(result, addr2);
-        Ok(())
-    })
-    .unwrap();
-}
-
-/// Regression test for https://github.com/0xMiden/compiler/issues/872
-///
-/// Previously, compilation could panic during stack manipulation with:
-/// `invalid stack index: only the first 16 elements on the stack are directly accessible, got 16`.
-#[test]
-fn test_invalid_stack_index_16_issue_872() {
-    let main_fn = r#"
-        (a0: Felt, a1: Felt, a2: Felt, a3: Felt, a4: Felt, a5: Felt, a6: Felt, a7: Felt,
-         a8: Felt, a9: Felt, a10: Felt, a11: Felt, a12: Felt, a13: Felt, a14: Felt, a15: Felt) -> Felt {
-            // Keep locals live across the call which are used only after the call, so that the 16
-            // call arguments are not at the top of the operand stack at call time.
-            let post = a0 + miden_stdlib_sys::felt!(1);
-
-            let res = callee_16(a0, a1, a2, a3, a4, a5, a6, a7, a8, a9, a10, a11, a12, a13, a14, a15);
-
-            // Use all post-call locals to prevent DCE.
-            res + post
-        }
-
-        #[inline(never)]
-        fn callee_16(
-            a0: Felt, a1: Felt, a2: Felt, a3: Felt, a4: Felt, a5: Felt, a6: Felt, a7: Felt,
-            a8: Felt, a9: Felt, a10: Felt, a11: Felt, a12: Felt, a13: Felt, a14: Felt, a15: Felt,
-        ) -> Felt {
-            a0 + a1 + a2 + a3 + a4 + a5 + a6 + a7 + a8 + a9 + a10 + a11 + a12 + a13 + a14 + a15
-        }
-    "#;
-
-    setup::enable_compiler_instrumentation();
-    let config = WasmTranslationConfig::default();
-    let mut test =
-        CompilerTest::rust_fn_body_with_stdlib_sys("movup_16_issue_831", main_fn, config, []);
-
-    let package = test.compile_package();
-
-    // This should execute and return the expected value.
-    let args: [Felt; 16] = [
-        Felt::from(1u32),
-        Felt::from(2u32),
-        Felt::from(3u32),
-        Felt::from(4u32),
-        Felt::from(5u32),
-        Felt::from(6u32),
-        Felt::from(7u32),
-        Felt::from(8u32),
-        Felt::from(9u32),
-        Felt::from(10u32),
-        Felt::from(11u32),
-        Felt::from(12u32),
-        Felt::from(13u32),
-        Felt::from(14u32),
-        Felt::from(15u32),
-        Felt::from(16u32),
-    ];
-
-    let expected = (1u32..=16u32).fold(Felt::ZERO, |acc, x| acc + Felt::from(x)) + Felt::from(2u32);
-
-    eval_package::<Felt, _, _>(&package, [], &args, &test.session, |trace| {
-        let res: Felt = trace.parse_result().unwrap();
-        assert_eq!(res, expected);
-        Ok(())
-    })
-    .unwrap();
-}
-
-/// Regression test for the case of 4 words + 1 felt (17 felts) function args the words are passed by pointer.
-#[test]
-fn test_invalid_stack_index_4_word_1_felt_args() {
-    let main_fn = r#"
-        (
-            w0_0: Felt,
-            w0_1: Felt,
-            w0_2: Felt,
-            w0_3: Felt,
-            w1_0: Felt,
-            w1_1: Felt,
-            w1_2: Felt,
-            w1_3: Felt,
-            w2_0: Felt,
-            w2_1: Felt,
-            w2_2: Felt,
-            w2_3: Felt,
-            w3_0: Felt,
-            w3_1: Felt,
-            w3_2: Felt,
-            w3_3: Felt,
-        ) -> Felt {
-            let w0 = Word::new([w0_0, w0_1, w0_2, w0_3]);
-            let w1 = Word::new([w1_0, w1_1, w1_2, w1_3]);
-            let w2 = Word::new([w2_0, w2_1, w2_2, w2_3]);
-            let w3 = Word::new([w3_0, w3_1, w3_2, w3_3]);
-
-            // Keep locals live across the call which are used only after the call, so that the
-            // call arguments are not at the top of the operand stack at call time.
-            let post = w0[0] + miden_stdlib_sys::felt!(1);
-
-            let extra = w0[1];
-            let res = callee_5(w0, w1, w2, w3, extra);
-
-            // Use all post-call locals to prevent DCE.
-            res + post
-        }
-
-        #[inline(never)]
-        fn callee_5(w0: Word, w1: Word, w2: Word, w3: Word, extra: Felt) -> Felt {
-            w0[0] + w0[1] + w0[2] + w0[3] +
-            w1[0] + w1[1] + w1[2] + w1[3] +
-            w2[0] + w2[1] + w2[2] + w2[3] +
-            w3[0] + w3[1] + w3[2] + w3[3] +
-            extra
-        }
-    "#;
-
-    setup::enable_compiler_instrumentation();
-    let config = WasmTranslationConfig::default();
-    let mut test = CompilerTest::rust_fn_body_with_stdlib_sys(
-        "movup_4_word_1_felt_args_invalid_stack_index",
-        main_fn,
-        config,
-        [],
-    );
-
-    let package = test.compile_package();
-
-    // This should execute and return the expected value.
-    let args: [Felt; 16] = [
-        // w0
-        Felt::from(1u32),
-        Felt::from(2u32),
-        Felt::from(3u32),
-        Felt::from(4u32),
-        // w1
-        Felt::from(5u32),
-        Felt::from(6u32),
-        Felt::from(7u32),
-        Felt::from(8u32),
-        // w2
-        Felt::from(9u32),
-        Felt::from(10u32),
-        Felt::from(11u32),
-        Felt::from(12u32),
-        // w3
-        Felt::from(13u32),
-        Felt::from(14u32),
-        Felt::from(15u32),
-        Felt::from(16u32),
-    ];
-
-    // Expected:
-    // - callee_5 sums 1..=16 and adds `extra` (w0[1] == 2)
-    // - main adds `post` (w0[0] + 1 == 2)
-    let expected = (1u32..=16u32).fold(Felt::ZERO, |acc, x| acc + Felt::from(x)) + Felt::from(4u32);
-
-    eval_package::<Felt, _, _>(&package, [], &args, &test.session, |trace| {
-        let res: Felt = trace.parse_result().unwrap();
-        assert_eq!(res, expected);
-        Ok(())
-    })
-    .unwrap();
-}
-
-/// Regression test for https://github.com/0xMiden/compiler/issues/811
-///
-/// This reproduces a bug (byte-by-byte copying) in `memory.copy` lowering used by `Vec` reallocation.
-#[test]
-fn test_vec_realloc_copies_data_issue_811() {
-    let main_fn = r#"() -> Felt {
-        extern crate alloc;
-        use alloc::vec::Vec;
-
-        // Create a Vec with a tiny capacity to make growth (and thus reallocation) likely.
-        let mut v: Vec<Felt> = Vec::with_capacity(1);
-
-        v.push(felt!(11111));
-        let mut last_ptr = v.as_ptr() as u32;
-        let mut moves: u32 = 0;
-
-        v.push(felt!(22222));
-        let ptr = v.as_ptr() as u32;
-        if ptr != last_ptr {
-            moves += 1;
-            last_ptr = ptr;
-        }
-
-        v.push(felt!(33333));
-        let ptr = v.as_ptr() as u32;
-        if ptr != last_ptr {
-            moves += 1;
-            last_ptr = ptr;
-        }
-
-        v.push(felt!(44444));
-        let ptr = v.as_ptr() as u32;
-        if ptr != last_ptr {
-            moves += 1;
-            last_ptr = ptr;
-        }
-
-        v.push(felt!(55555));
-        let ptr = v.as_ptr() as u32;
-        if ptr != last_ptr {
-            moves += 1;
-            last_ptr = ptr;
-        }
-
-        // Sum all elements - if realloc doesn't copy, the first 4 elements will be garbage.
-        let sum = v[0] + v[1] + v[2] + v[3] + v[4];
-        if moves >= 2 { sum } else { felt!(0) }
-    }"#;
-
-    setup::enable_compiler_instrumentation();
-    let config = WasmTranslationConfig::default();
-    let mut test =
-        CompilerTest::rust_fn_body_with_stdlib_sys("vec_realloc_copies_data", main_fn, config, []);
-
-    let package = test.compile_package();
-    let args: [Felt; 0] = [];
-
-    eval_package::<Felt, _, _>(&package, [], &args, &test.session, |trace| {
-        let result: u64 = trace.parse_result::<Felt>().unwrap().as_canonical_u64();
-        assert_eq!(result, 166_665, "Vec reallocation failed to copy existing elements");
-        Ok(())
-    })
-    .unwrap();
-}
-
-#[ignore = "too fragile (depends on mem addrs), this bug is also covered by the test_hmerge test"]
-#[test]
-fn test_func_arg_order() {
-    // This test reproduces the "swapped and frozen" function arguments issue
-    // https://github.com/0xMiden/compiler/pull/576 discovered while working on hmerge VM op
-    // The issue manifests in "intrinsic" function parameters being in the wrong order
-    // (see assert_eq before the call and inside the function)
-    // on the stack AND the their order is not changing when the parameters are
-    // swapped at the call site (see expect_masm with the same file name, i.e. the MASM
-    // do not change when she parameters are swapped).
-    fn main_fn_template(digest_ptr_name: &str, result_ptr_name: &str) -> String {
-        format!(
-            r#"
-    (f0: miden_stdlib_sys::Felt, f1: miden_stdlib_sys::Felt, f2: miden_stdlib_sys::Felt, f3: miden_stdlib_sys::Felt, f4: miden_stdlib_sys::Felt, f5: miden_stdlib_sys::Felt, f6: miden_stdlib_sys::Felt, f7: miden_stdlib_sys::Felt) -> miden_stdlib_sys::Felt {{
-        let digest1 = miden_stdlib_sys::Digest::new([f0, f1, f2, f3]);
-        let digest2 = miden_stdlib_sys::Digest::new([f4, f5, f6, f7]);
-        let digests = [digest1, digest2];
-        let res = merge(digests);
-        res.inner[0]
-    }}
-
-    #[inline]
-    pub fn merge(digests: [Digest; 2]) -> Digest {{
-        unsafe {{
-            let digests_ptr = digests.as_ptr().addr() as u32;
-            assert_eq(Felt::from_u32(digests_ptr as u32), Felt::from_u32(1048528));
-
-            let mut ret_area = ::core::mem::MaybeUninit::<Word>::uninit();
-            let result_ptr = ret_area.as_mut_ptr().addr() as u32;
-            assert_eq(Felt::from_u32(result_ptr as u32), Felt::from_u32(1048560));
-
-            intrinsic({digest_ptr_name} as *const Felt, {result_ptr_name} as *mut Felt);
-
-            Digest::from_word(ret_area.assume_init())
-        }}
-    }}
-
-    #[unsafe(no_mangle)]
-    fn intrinsic(digests_ptr: *const Felt, result_ptr: *mut Felt) {{
-        // see assert_eq above, before the call
-        assert_eq(Felt::from_u32(digests_ptr as u32), Felt::from_u32(1048528));
-        // see assert_eq above, before the call
-        assert_eq(Felt::from_u32(result_ptr as u32), Felt::from_u32(1048560));
-    }}
-        "#
-        )
-    }
-
-    let config = WasmTranslationConfig::default();
-    let main_fn_correct = main_fn_template("digests_ptr", "result_ptr");
-    let mut test = CompilerTest::rust_fn_body_with_stdlib_sys(
-        "func_arg_order",
-        &main_fn_correct,
-        config.clone(),
-        [],
-    );
-
-    let args = [
-        Felt::ZERO,
-        Felt::ZERO,
-        Felt::ZERO,
-        Felt::ZERO,
-        Felt::ZERO,
-        Felt::ZERO,
-        Felt::ZERO,
-        Felt::ZERO,
-    ];
-
-    eval_package::<Felt, _, _>(&test.compile_package(), [], &args, &test.session, |trace| Ok(()))
-        .unwrap();
-}
