+++ tests/integration/src/rust_pipeline/functional/types/array.rs
@@ -1,0 +1,6 @@
+use crate::CompilerTest;
+
+#[test]
+fn test_array() {
+    let _ = CompilerTest::rust_source_program(include_str!("types_src/array.rs")).compile_package();
+}
