+++ tests/integration/src/rust_pipeline/property/abi_transform/stdlib/mod.rs
@@ -1,0 +1,5 @@
+mod hash_elements;
+mod hash_words;
+mod pipe_double_words_to_memory;
+mod pipe_words_to_memory;
+mod vec_alloc_vec;
